+++ packages/spotlight/src/container.js
@@ -7,6 +7,7 @@
 
 import and from 'ramda/src/and';
 import concat from 'ramda/src/concat';
+import {isWindowReady} from '@enact/core/snapshot';
 import {coerceArray} from '@enact/core/util';
 import intersection from 'ramda/src/intersection';
 import last from 'ramda/src/last';
@@ -88,17 +89,14 @@ let GlobalConfig = {
  * @private
  */
 const querySelector = (node, includeSelector, excludeSelector) => {
-	const include = Array.prototype.slice.call(node.querySelectorAll(includeSelector));
+	const include = new Set(node.querySelectorAll(includeSelector));
 	const exclude = node.querySelectorAll(excludeSelector);
 
 	for (let i = 0; i < exclude.length; i++) {
-		const index = include.indexOf(exclude.item(i));
-		if (index >= 0) {
-			include.splice(index, 1);
-		}
+		include.delete(exclude[i]);
 	}
 
-	return include;
+	return Array.from(include);
 };
 
 /**
@@ -261,7 +259,13 @@ const getContainerNode = (containerId) => {
  * @private
  */
 const navigableFilter = (node, containerId) => {
+	const nodeStyle = node && isWindowReady() && window.getComputedStyle(node);
 	const config = getContainerConfig(containerId);
+
+	if (!nodeStyle || nodeStyle.display === 'none' || nodeStyle.visibility === 'hidden') {
+		return false;
+	}
+
 	if (config && typeof config.navigableFilter === 'function') {
 		if (config.navigableFilter(node, containerId) === false) {
 			return false;
@@ -623,9 +627,8 @@ const getAllContainerIds = () => {
 /**
  * Returns the default focus element for a container
  *
- * @param   {String}                             containerId        ID of container
- * @param   {('last-focused'|'default-element')} [preferredEnterTo] Prefer the given enterTo
- *                                                                  configuration
+ * @param   {String}                                       containerId        Container ID
+ * @param   {('last-focused'|'default-element'|'topmost')} [preferredEnterTo] Prefer the given enterTo configuration
  *
  * @returns {Node|null}                 Default focus element
  * @memberof spotlight/container
@@ -722,9 +725,8 @@ function setContainerLastFocusedElement (node, containerIds) {
  * return that element as the only element in an array. If that fails or if navigation is not
  * restricted, it will return an array of all possible navigable nodes.
  *
- * @param   {String}                             containerId        Container ID
- * @param   {('last-focused'|'default-element')} [preferredEnterTo] Prefer the given enterTo
- *                                                                  configuration
+ * @param   {String}                                       containerId        Container ID
+ * @param   {('last-focused'|'default-element'|'topmost')} [preferredEnterTo] Prefer the given enterTo configuration
  *
  * @returns {Node[]}             Navigable elements within container
  * @memberof spotlight/container
@@ -769,6 +771,10 @@ function getContainerNavigableElements (containerId, preferredEnterTo) {
 
 				return contains(containerRect, getRect(element));
 			});
+
+			if (next && preferredEnterTo === 'topmost') {
+				next.sort((a, b) => (getRect(a).top - getRect(b).top));
+			}
 		}
 
 		// otherwise, return all spottables within the container
@@ -784,9 +790,8 @@ function getContainerNavigableElements (containerId, preferredEnterTo) {
  * Determines the preferred focus target, traversing any sub-containers as necessary, for the given
  * container.
  *
- * @param   {String}                             containerId        ID of container
- * @param   {('last-focused'|'default-element')} [preferredEnterTo] Prefer the given enterTo
- *                                                                  configuration
+ * @param   {String}                                       containerId        Container ID
+ * @param   {('last-focused'|'default-element'|'topmost')} [preferredEnterTo] Prefer the given enterTo configuration
  *
  * @returns {Node}                 Preferred target as either a DOM node or container-id
  * @memberof spotlight/container