+++ packages/ui/VirtualList/UiVirtualListBase.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/no-this-in-sfc */
 import classNames from 'classnames';
 import {forward} from '@enact/core/handle';
 import {platform} from '@enact/core/platform';
@@ -111,20 +112,12 @@ const VirtualListBaseFactory = (type) => {
 			}),
 
 			/**
-			 * Activates the component for voice control.
+			 * An object with properties to be passed to the container DOM.
 			 *
-			 * @type {Boolean}
-			 * @public
+			 * @type {Object}
+			 * @private
 			 */
-			'data-webos-voice-focused': PropTypes.bool,
-
-			/**
-			 * The voice control group label.
-			 *
-			 * @type {String}
-			 * @public
-			 */
-			'data-webos-voice-group-label': PropTypes.string,
+			containerProps: PropTypes.object,
 
 			/**
 			 * The number of items of data the list contains.
@@ -224,6 +217,7 @@ const VirtualListBaseFactory = (type) => {
 			cbScrollTo: nop,
 			dataSize: 0,
 			direction: 'vertical',
+			itemsRenderer: nop, // eslint-disable-line react/default-props-match-prop-types
 			overhang: 3,
 			pageScroll: false,
 			spacing: 0
@@ -276,7 +270,6 @@ const VirtualListBaseFactory = (type) => {
 		componentDidMount () {
 			if (!this.props.clientSize) {
 				this.calculateMetrics(this.props);
-				// eslint-disable-next-line react/no-did-mount-set-state
 				this.setState(this.getStatesAndUpdateBounds(this.props));
 			} else {
 				this.emitUpdateItems();
@@ -354,7 +347,6 @@ const VirtualListBaseFactory = (type) => {
 				const {x, y} = this.getXY(this.scrollPosition, 0);
 
 				this.calculateMetrics(this.props);
-				// eslint-disable-next-line react/no-did-update-set-state
 				this.setState(this.getStatesAndUpdateBounds(this.props));
 				this.setContainerSize();
 
@@ -370,7 +362,6 @@ const VirtualListBaseFactory = (type) => {
 				deferScrollTo = true;
 			} else if (this.hasDataSizeChanged) {
 				const newState = this.getStatesAndUpdateBounds(this.props, this.state.firstIndex);
-				// eslint-disable-next-line react/no-did-update-set-state
 				this.setState(newState);
 				this.setContainerSize();
 
@@ -1132,7 +1123,7 @@ const VirtualListBaseFactory = (type) => {
 
 		render () {
 			const
-				{className, 'data-webos-voice-focused': voiceFocused, 'data-webos-voice-group-label': voiceGroupLabel, itemsRenderer, style, ...rest} = this.props,
+				{className, containerProps, itemsRenderer, style, ...rest} = this.props,
 				{cc, itemContainerRef, primary} = this,
 				containerClasses = this.getContainerClasses(className),
 				contentClasses = this.getContentClasses();
@@ -1161,7 +1152,7 @@ const VirtualListBaseFactory = (type) => {
 			}
 
 			return (
-				<div className={containerClasses} data-webos-voice-focused={voiceFocused} data-webos-voice-group-label={voiceGroupLabel} ref={this.containerRef} style={style}>
+				<div className={containerClasses} {...containerProps} ref={this.containerRef} style={style}>
 					<div {...rest} className={contentClasses} ref={this.contentRef}>
 						{itemsRenderer({cc, itemContainerRef, primary})}
 					</div>