+++ packages/spotlight/CHANGELOG.md
@@ -2,7 +2,21 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
-## [unreleased]
+## [4.5.0-alpha.2] - 2022-05-09
+
+### Changed
+
+- `spotlight` to not focus on an invisible element
+
+## [4.5.0-alpha.1] - 2022-04-15
+
+No significant changes.
+
+## [4.1.4] - 2022-03-24
+
+No significant changes.
+
+## [4.1.3] - 2022-03-07
 
 - Updated to use `forwardCustom` and add `type` when forwarding custom events